♻️ address review: relocate tests, trim docs, fix changelog
Move the multi-qubit noise validation into test_digital_tjm.py to match
src/, shorten the circuit_observables note, and align the changelog entry
with PR #529.
